pyproject.toml
[tool.poetry]
name = "flask-ory-auth"
version = "0.1.0"
description = "Useful middlewares and clients to use Ory products in flask App"
authors = ["Andrew Minkin <minkin.andrew@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.7"
requests = "^2.28.0"
requests-oauthlib = "^1.3.1"
Flask = "^2.1.2"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
autoflake = "^1.4"
pre-commit = "^2.19.0"
black = "^22.3.0"
yesqa = "^1.3.0"
reorder-python-imports = "^3.1.0"
pyupgrade = "^2.34.0"
pytest = "^7.1.2"
pytest-flask = "^1.2.0"
responses = "^0.21.0"
tox = "^3.25.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
